+++ src/mvvm/awtk/binding_context_awtk.c
@@ -188,7 +188,7 @@ static ret_t widget_set_prop_if_diff(widget_t* widget, const char* name, const v
                                      bool_t set_force) {
   value_t old;
   return_value_if_fail(widget != NULL && widget->vt != NULL, RET_BAD_PARAMS);
-  
+
   if (!set_force) {
     if (widget->vt->inputable) {
       if (tk_str_eq(name, WIDGET_PROP_TEXT) || tk_str_eq(name, WIDGET_PROP_VALUE)) {
@@ -941,7 +941,7 @@ static ret_t binding_context_awtk_update_data_of_widget(binding_context_t* ctx, 
   binding_rule_t* rule = NULL;
   slist_node_t* iter = NULL;
   return_value_if_fail(ctx != NULL, RET_BAD_PARAMS);
-  
+
   iter = ctx->data_bindings.first;
 
   while (iter != NULL) {
@@ -1025,47 +1025,12 @@ static ret_t widget_visit_dynamic_binding_update_to_view(void* ctx, const void* 
 static ret_t binding_context_awtk_update_to_view_sync(binding_context_t* ctx) {
   return_value_if_fail(ctx != NULL, RET_BAD_PARAMS);
 
-  ctx->updating_view = TRUE;
-
   slist_foreach(&(ctx->dynamic_bindings), widget_visit_dynamic_binding_update_to_view, FALSE);
   slist_foreach(&(ctx->data_bindings), widget_visit_data_binding_update_to_view, NULL);
   slist_foreach(&(ctx->command_bindings), widget_visit_command_binding_update_to_view, NULL);
   widget_invalidate_force(WIDGET(ctx->widget), NULL);
 
-  ctx->updating_view = FALSE;
   ctx->updating_view_by_ui = FALSE;
-
-  return RET_OK;
-}
-
-static ret_t idle_update_to_view(const idle_info_t* info) {
-  binding_context_t* ctx = BINDING_CONTEXT(info->ctx);
-  return_value_if_fail(ctx != NULL, RET_BAD_PARAMS);
-
-  if (ctx->bound && !mvvm_awtk_is_quited()) {
-    binding_context_awtk_update_to_view_sync(ctx);
-  } else {
-    log_debug("binding context is unbound, or mvvm awtk is quited.\n");
-  }
-
-  // update 结束后再复位，以便 binding_context_update_to_view 判断是否循环调用
-  ctx->update_view_idle_id = TK_INVALID_ID;
-
-  return RET_REMOVE;
-}
-
-static ret_t binding_context_awtk_update_to_view(binding_context_t* ctx) {
-  return_value_if_fail(ctx != NULL && ctx->view_model != NULL, RET_BAD_PARAMS);
-
-  if (ctx->bound || ctx->updating_view) {
-    if (ctx->update_view_idle_id == TK_INVALID_ID) {
-      ctx->update_view_idle_id = idle_add(idle_update_to_view, ctx);
-    }
-  } else {
-    if (ctx->widget != NULL) {
-      binding_context_awtk_update_to_view_sync(ctx);
-    }
-  }
 
   return RET_OK;
 }
@@ -1096,7 +1061,7 @@ static ret_t binding_context_awtk_remove_all_rebind_idle(binding_context_t* ctx)
   tk_compare_t compare = (tk_compare_t)binding_context_awtk_compare_items_binding;
   slist_t* bindings = NULL;
   return_value_if_fail(ctx != NULL, RET_BAD_PARAMS);
-  
+
   bindings = &(ctx->dynamic_bindings);
 
   darray_init(&matched, 1, NULL, NULL);
@@ -1157,7 +1122,7 @@ static ret_t binding_context_awtk_destroy(binding_context_t* ctx) {
 
 static const binding_context_vtable_t s_binding_context_vtable = {
     .destroy = binding_context_awtk_destroy,
-    .update_to_view = binding_context_awtk_update_to_view,
+    .update_to_view = binding_context_awtk_update_to_view_sync,
     .update_to_model = binding_context_awtk_update_to_model,
     .exec = binding_context_awtk_exec,
     .can_exec = binding_context_awtk_can_exec,