+++ src/mvvm/base/binding_context.h
@@ -23,6 +23,7 @@
 #define TK_BINDING_CONTEXT_H
 
 #include "tkc/slist.h"
+#include "tkc/idle_info.h"
 #include "mvvm/base/mvvm_types_def.h"
 #include "mvvm/base/view_model.h"
 #include "mvvm/base/binding_rule.h"
@@ -408,6 +409,9 @@ ret_t binding_context_get_prop_by_rule(binding_context_t* ctx, binding_rule_t* r
  */
 ret_t binding_context_set_prop_by_rule(binding_context_t* ctx, binding_rule_t* rule,
                                        const char* name, const value_t* v);
+
+/*internal use*/
+ret_t binding_context_idle_update_to_view(const idle_info_t* info);
 
 #define BINDING_CONTEXT(ctx) ((binding_context_t*)(ctx))